Remove old accessor
